Clean up idle streams in UDP syslog target
The UDP transport in the syslog target creates a goroutine for each
remote address to handle multiple packets coming from the same address.
However, these goroutines were not cleaned up.

This commit adds an idle timeout for these goroutines waiting for
packets to be processed in order to prevent goroutine leaking.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>